chore: remove secret GH_TOKEN_FOR_UPDATES. Closes #51
The secret GH_TOKEN_FOR_UPDATES was used to run the CI actions
on the PR opened by the update-flake-lock action.  This is not
possible anymore because we moved the repository to an organization.

The drawback of this is that the CI has to be manually invoked on the
PR following the instructions at https://github.com/DeterminateSystems/update-flake-lock?tab=readme-ov-file#without-a-personal-authentication-token